Start moving datamanagement to plugin. Cleanup readers/writers in vtkXXXXX files;
